api/requirements_test.txt
-r requirements.txt
pytest-aiohttp==0.3.0
pycodestyle
coverage
flake8
mypy
